Refactor Microservice Module
