+++ application/services/telegraph/src/config/services/microservcies-config/microservice-module-options.ts
@@ -1,0 +1,13 @@
+import {ASYNC_OPTIONS_TYPE} from '@looport/nest-microservice'
+
+import {ConfigModule} from '@/config/config.module'
+import {MicroserviceConfigService} from '@/config/services/microservcies-config/microservice-config.service'
+
+export const getMicroserviceModuleAsyncOptions =
+  (): typeof ASYNC_OPTIONS_TYPE => ({
+    imports: [ConfigModule],
+    inject: [MicroserviceConfigService],
+    isGlobal: true,
+    useFactory: (configService: MicroserviceConfigService) =>
+      configService.getNATSConfig(),
+  })
