+++ application/services/telegraph/src/message/dto/message/message-serializer.ts
@@ -1,6 +1,6 @@
 import {instanceToPlain, plainToInstance} from 'class-transformer'
 
-import {MessageDto} from '@/messages/dto/message/message.dto'
+import {MessageDto} from '@/message/dto/message/message.dto'
 import {MessageEntity} from '@/storage/entities/message.entity'
 
 export const serializeMessage = (message: MessageEntity) =>
