+++ application/services/telegraph/src/room/controllers/rooms/room.controller.test.ts
@@ -6,8 +6,8 @@ import {Test, TestingModule} from '@nestjs/testing'
 
 import {MessageEventService} from '@/event/services/message-event.service'
 import {createMessageEventMock} from '@/event/services/message-event.service.mock'
-import {MessageService} from '@/messages/services/messages/message.service'
-import {createMessageServiceMock} from '@/messages/services/messages/message.service.mock'
+import {MessageService} from '@/message/services/messages/message.service'
+import {createMessageServiceMock} from '@/message/services/messages/message.service.mock'
 import {RoomService} from '@/room/services/rooms/room.service'
 import {createRoomMockService} from '@/room/services/rooms/room.service.mock'
 import {messageMock} from '@/storage/repositories/message/message.repository.mock'
