+++ application/services/telegraph/src/room/controllers/rooms/room.controller.ts
@@ -5,9 +5,9 @@ import {fromEvent, map, skipWhile} from 'rxjs'
 
 import {MessageEventService} from '@/event/services/message-event.service'
 import {createMessageCreatedSubject} from '@/event/services/message-event.service.lib'
-import {CreateMessageDto} from '@/messages/dto/create-message.dto'
-import {MessageDto} from '@/messages/dto/message/message.dto'
-import {MessageService} from '@/messages/services/messages/message.service'
+import {CreateMessageDto} from '@/message/dto/create-message.dto'
+import {MessageDto} from '@/message/dto/message/message.dto'
+import {MessageService} from '@/message/services/messages/message.service'
 import {RoomService} from '@/room/services/rooms/room.service'
 import {MessageEntity} from '@/storage/entities/message.entity'
 import {RoomEntity} from '@/storage/entities/room.entity'
