+++ application/services/telegraph/src/room/room.module.ts
@@ -1,7 +1,7 @@
 import {Module} from '@nestjs/common'
 
 import {EventModule} from '@/event/event.module'
-import {MessageModule} from '@/messages/message.module'
+import {MessageModule} from '@/message/message.module'
 import {StorageModule} from '@/storage/storage.module'
 
 import {RoomController} from './controllers/rooms/room.controller'
