+++ application/services/telegraph/src/test/room/room.e2e-test.ts
@@ -8,8 +8,7 @@ import request from 'supertest'
 
 import {AppModule} from '@/app/app.module'
 import {createMessageCreatedSubject} from '@/event/services/message-event.service.lib'
-import {serializeMessage} from '@/messages/dto/message/message-serializer'
-import {MessageDto} from '@/messages/dto/message/message.dto'
+import {serializeMessage} from '@/message/dto/message/message-serializer'
 import {RoomService} from '@/room/services/rooms/room.service'
 import {MessageRepository} from '@/storage/repositories/message/message.repository'
 import {messageMock} from '@/storage/repositories/message/message.repository.mock'
